+++ NEWS.md
@@ -12,6 +12,7 @@
 * `read_netCDF_as_array()` gains arguments `"time_name"` and `"vertical_name"`
   with previous hard-coded values as defaults ("time" and "vertical")
   (#14; @dschlaep)
+* New `get_nc_type()` identifies a suitable netCDF data type from an R object.
 
 
 # rSW2st v0.1.0